refactor(core): split canvas DOM module by concern
Separate leaf DOM helpers (color/DPR/clip), mark drawers + drawBatch, and
stratum panel routing so paint paths and viewport recipes can change
independently. Preserve the @ggsvelte/core/dom public barrel.

Split the monolithic canvas test file by routing/focus/segments, share
fixtures, and add a barrel + DOM-helper smoke suite.
