+++ packages/core/src/dom/canvas.ts
@@ -1,6 +1,6 @@
 /// <reference lib="dom" />
 /**
- * Canvas batch renderers (graduated from spike 0006).
+ * Canvas stratum orchestration (graduated from spike 0006).
  *
  * Draws GeometryBatch kinds with THE SAME style semantics as the SVG
  * renderers (render-svg.ts / <Batch>): null styling values mean the theme
@@ -16,535 +16,27 @@
  * CSS box pinned via style, ONE absolute setTransform so all drawing code
  * stays in CSS px and resize re-runs cannot compound.
  *
- * Glyph (text) batches are NOT drawn here: text always renders as SVG
+ * Glyph (text) batches are NOT drawn on canvas: text always renders as SVG
  * (plan: "axes/grids/legends/text always SVG"); planStrata routes them.
+ *
+ * Module split: leaf DOM helpers in `canvas-dom.ts`, mark drawers in
+ * `canvas-marks.ts`, panel routing + focus presentation here.
  */
-import type { GeometryBatch, PathsBatch, PointsBatch, Scene, SegmentsBatch } from "../scene.js";
-import type { BatchInteractionMask } from "../interaction-mask.js";
-import type { ThemeTokens } from "../theme.js";
-import { themeVar } from "../theme.js";
+import type { GeometryBatch, Scene } from "../scene.js";
+import type { ColorResolver } from "./canvas-dom.js";
+import { drawClippedToPanel } from "./canvas-dom.js";
+import type { CanvasFocusPresentation } from "./canvas-marks.js";
+import {
+  batchPrimitiveCount,
+  drawBatch,
+  drawBatchSubset,
+  maskIsAllFocused,
+} from "./canvas-marks.js";
 
-/** Resolve a scene color expression to a concrete canvas-usable color. */
-export type ColorResolver = (color: string) => string;
-
-/**
- * Renderer-neutral focus membership for one geometry batch. Values address
- * point/rect/segment/glyph primitives, or path subpaths (not path vertices).
- * Zero means muted and a non-zero value means focused.
- */
-export type PrimitiveFocusMask = ArrayLike<number> | BatchInteractionMask;
-
-/** Optional focus presentation aligned to the `batches` passed to drawStratum. */
-export interface CanvasFocusPresentation {
-  readonly focusMasks: readonly (PrimitiveFocusMask | null | undefined)[];
-  /** Defaults to the scene theme's interactionMuted token. */
-  readonly mutedAlpha?: number;
-}
-
-function maskIncludes(mask: PrimitiveFocusMask, index: number): boolean {
-  if ("isFocused" in mask) return mask.isFocused(index);
-  const value = mask[index];
-  return value !== undefined && value !== 0;
-}
-
-const VAR_RE = /^var\((--[\w-]+)\s*,\s*(.+)\)$/;
-
-/**
- * Build a resolver over an element's computed style: `var(--gg-*, fallback)`
- * reads the custom property (falling back to the embedded token), and
- * `currentColor` reads the computed `color`. Resolutions are cached; build
- * a fresh resolver per draw so theme flips re-resolve.
- */
-export function cssColorResolver(el: Element): ColorResolver {
-  const cache = new Map<string, string>();
-  let computed: CSSStyleDeclaration | null = null;
-  const style = () => (computed ??= getComputedStyle(el));
-  return (color: string): string => {
-    const cached = cache.get(color);
-    if (cached !== undefined) return cached;
-    let out = color;
-    const match = VAR_RE.exec(out);
-    if (match !== null) {
-      const value = style().getPropertyValue(match[1]!).trim();
-      out = value === "" ? match[2]! : value;
-    }
-    if (out === "currentColor") out = style().color;
-    cache.set(color, out);
-    return out;
-  };
-}
-
-/**
- * Size a canvas for a device pixel ratio (decision 0006 recipe): rounded
- * integer backing store, pinned CSS box, absolute dpr transform.
- */
-export function sizeCanvasForDpr(
-  canvas: HTMLCanvasElement,
-  ctx: CanvasRenderingContext2D,
-  cssWidth: number,
-  cssHeight: number,
-  dpr: number,
-): void {
-  canvas.width = Math.round(cssWidth * dpr);
-  canvas.height = Math.round(cssHeight * dpr);
-  canvas.style.width = `${cssWidth}px`;
-  canvas.style.height = `${cssHeight}px`;
-  ctx.setTransform(dpr, 0, 0, dpr, 0, 0);
-}
-
-/**
- * Panel clipping recipe (decision 0006): save / beginPath / rect / clip /
- * draw / restore. beginPath is load-bearing — clip() uses the current path.
- */
-export function drawClippedToPanel(
-  ctx: CanvasRenderingContext2D,
-  panel: { x: number; y: number; width: number; height: number; clip?: boolean },
-  draw: (ctx: CanvasRenderingContext2D) => void,
-): void {
-  if (panel.clip === false) {
-    draw(ctx);
-    return;
-  }
-  ctx.save();
-  ctx.beginPath();
-  ctx.rect(panel.x, panel.y, panel.width, panel.height);
-  ctx.clip();
-  draw(ctx);
-  ctx.restore();
-}
-
-function tracePoint(ctx: CanvasRenderingContext2D, batch: PointsBatch, j: number): void {
-  const x = batch.positions[j * 2]!;
-  const y = batch.positions[j * 2 + 1]!;
-  const size = batch.size;
-  switch (batch.shape) {
-    case "square":
-      ctx.rect(x - size, y - size, size * 2, size * 2);
-      break;
-    case "triangle":
-      // Same proportions as the SVG renderer's triangle path.
-      ctx.moveTo(x, y - size * 1.2);
-      ctx.lineTo(x + size * 1.1, y + size * 0.9);
-      ctx.lineTo(x - size * 1.1, y + size * 0.9);
-      ctx.closePath();
-      break;
-    default:
-      ctx.moveTo(x + size, y);
-      ctx.arc(x, y, size, 0, Math.PI * 2);
-  }
-}
-
-function drawPoints(
-  ctx: CanvasRenderingContext2D,
-  batch: PointsBatch,
-  theme: ThemeTokens,
-  resolve: ColorResolver,
-): void {
-  const themeInk = resolve(themeVar("ink", theme));
-  const n = batch.rowIndex.length;
-  if (batch.colors === undefined) {
-    // Single fill: one path for the whole batch (the fast path that makes
-    // canvas worth it at high counts).
-    ctx.fillStyle = batch.fill === null ? themeInk : resolve(batch.fill);
-    ctx.beginPath();
-    for (let j = 0; j < n; j++) tracePoint(ctx, batch, j);
-    ctx.fill();
-    return;
-  }
-  // Per-mark colors: batch consecutive same-color runs.
-  let runStart = 0;
-  while (runStart < n) {
-    const color = batch.colors[runStart] ?? batch.fill ?? themeInk;
-    let runEnd = runStart + 1;
-    while (runEnd < n && (batch.colors[runEnd] ?? batch.fill ?? themeInk) === color) runEnd++;
-    ctx.fillStyle = resolve(color);
-    ctx.beginPath();
-    for (let j = runStart; j < runEnd; j++) tracePoint(ctx, batch, j);
-    ctx.fill();
-    runStart = runEnd;
-  }
-}
-
-function drawPointsSubset(
-  ctx: CanvasRenderingContext2D,
-  batch: PointsBatch,
-  theme: ThemeTokens,
-  resolve: ColorResolver,
-  mask: PrimitiveFocusMask,
-  focused: boolean,
-): void {
-  const includes = (index: number) => maskIncludes(mask, index) === focused;
-  const themeInk = resolve(themeVar("ink", theme));
-  const n = batch.rowIndex.length;
-  if (batch.colors === undefined) {
-    ctx.fillStyle = batch.fill === null ? themeInk : resolve(batch.fill);
-    ctx.beginPath();
-    let traced = false;
-    for (let j = 0; j < n; j++) {
-      if (!includes(j)) continue;
-      tracePoint(ctx, batch, j);
-      traced = true;
-    }
-    if (traced) ctx.fill();
-    return;
-  }
-  // Interactive masks must not turn alternating categorical colors into one
-  // beginPath/fill pair per point. For the normal small categorical case,
-  // batch by first-seen color while retaining the focused-over-muted pass
-  // boundary. Fall back to contiguous runs for high-cardinality colors.
-  const uniqueColors: string[] = [];
-  const seenColors = new Set<string>();
-  for (let j = 0; j < n && uniqueColors.length <= 64; j++) {
-    const color = batch.colors[j] ?? batch.fill ?? themeInk;
-    if (seenColors.has(color)) continue;
-    seenColors.add(color);
-    uniqueColors.push(color);
-  }
-  if (uniqueColors.length <= 64) {
-    for (const color of uniqueColors) {
-      ctx.fillStyle = resolve(color);
-      ctx.beginPath();
-      let traced = false;
-      for (let j = 0; j < n; j++) {
-        if ((batch.colors[j] ?? batch.fill ?? themeInk) !== color || !includes(j)) continue;
-        tracePoint(ctx, batch, j);
-        traced = true;
-      }
-      if (traced) ctx.fill();
-    }
-    return;
-  }
-  let runStart = 0;
-  while (runStart < n) {
-    const color = batch.colors[runStart] ?? batch.fill ?? themeInk;
-    let runEnd = runStart + 1;
-    while (runEnd < n && (batch.colors[runEnd] ?? batch.fill ?? themeInk) === color) runEnd++;
-    ctx.fillStyle = resolve(color);
-    ctx.beginPath();
-    let traced = false;
-    for (let j = runStart; j < runEnd; j++) {
-      if (!includes(j)) continue;
-      tracePoint(ctx, batch, j);
-      traced = true;
-    }
-    if (traced) ctx.fill();
-    runStart = runEnd;
-  }
-}
-
-function traceSubpath(
-  ctx: CanvasRenderingContext2D,
-  batch: PathsBatch,
-  start: number,
-  end: number,
-): void {
-  ctx.moveTo(batch.positions[start * 2]!, batch.positions[start * 2 + 1]!);
-  for (let i = start + 1; i < end; i++) {
-    const x = batch.positions[i * 2]!;
-    const y = batch.positions[i * 2 + 1]!;
-    if (batch.curve === "step") {
-      const prevX = batch.positions[(i - 1) * 2]!;
-      const prevY = batch.positions[(i - 1) * 2 + 1]!;
-      const mid = (prevX + x) / 2;
-      ctx.lineTo(mid, prevY);
-      ctx.lineTo(mid, y);
-    }
-    ctx.lineTo(x, y);
-  }
-  if (batch.closed === true) ctx.closePath();
-}
-
-function drawPaths(
-  ctx: CanvasRenderingContext2D,
-  batch: PathsBatch,
-  theme: ThemeTokens,
-  resolve: ColorResolver,
-): void {
-  const isArea = batch.fills !== undefined;
-  const subpaths = batch.pathOffsets.length - 1;
-  for (let s = 0; s < subpaths; s++) {
-    const start = batch.pathOffsets[s]!;
-    const end = batch.pathOffsets[s + 1]!;
-    if (end <= start) continue;
-    ctx.beginPath();
-    traceSubpath(ctx, batch, start, end);
-    if (isArea) {
-      ctx.fillStyle = resolve(batch.fills![s] ?? themeVar("accent", theme));
-      ctx.fill();
-    } else {
-      ctx.strokeStyle = resolve(batch.strokes[s] ?? themeVar("ink", theme));
-      ctx.lineWidth = batch.linewidth;
-      ctx.lineJoin = "round";
-      ctx.lineCap = "round";
-      ctx.stroke();
-    }
-  }
-}
-
-function drawPathsSubset(
-  ctx: CanvasRenderingContext2D,
-  batch: PathsBatch,
-  theme: ThemeTokens,
-  resolve: ColorResolver,
-  mask: PrimitiveFocusMask,
-  focused: boolean,
-): void {
-  const isArea = batch.fills !== undefined;
-  const subpaths = batch.pathOffsets.length - 1;
-  for (let s = 0; s < subpaths; s++) {
-    if (maskIncludes(mask, s) !== focused) continue;
-    const start = batch.pathOffsets[s]!;
-    const end = batch.pathOffsets[s + 1]!;
-    if (end <= start) continue;
-    ctx.beginPath();
-    traceSubpath(ctx, batch, start, end);
-    if (isArea) {
-      ctx.fillStyle = resolve(batch.fills![s] ?? themeVar("accent", theme));
-      ctx.fill();
-    } else {
-      ctx.strokeStyle = resolve(batch.strokes[s] ?? themeVar("ink", theme));
-      ctx.lineWidth = batch.linewidth;
-      ctx.lineJoin = "round";
-      ctx.lineCap = "round";
-      ctx.stroke();
-    }
-  }
-}
-
-/**
- * Trace one segment as a disconnected subpath (moveTo + lineTo).
- * Multiple segments share one path + stroke when strokeStyle matches —
- * same multi-subpath contract points use for mono/run fills (overlapping
- * antialiasing / alpha compositing differs from per-primitive stroke()).
- */
-function traceSegment(ctx: CanvasRenderingContext2D, batch: SegmentsBatch, j: number): void {
-  if (batch.renderPositions !== undefined && batch.renderPathOffsets !== undefined) {
-    const start = batch.renderPathOffsets[j]!;
-    const end = batch.renderPathOffsets[j + 1]!;
-    if (end <= start) return;
-    ctx.moveTo(batch.renderPositions[start * 2]!, batch.renderPositions[start * 2 + 1]!);
-    for (let vertex = start + 1; vertex < end; vertex++) {
-      ctx.lineTo(batch.renderPositions[vertex * 2]!, batch.renderPositions[vertex * 2 + 1]!);
-    }
-    return;
-  }
-  const o = j * 4;
-  ctx.moveTo(batch.segments[o]!, batch.segments[o + 1]!);
-  ctx.lineTo(batch.segments[o + 2]!, batch.segments[o + 3]!);
-}
-
-function segmentStrokeAt(
-  batch: SegmentsBatch,
-  j: number,
-  themeInk: string,
-  resolve: ColorResolver,
-): string {
-  const stroke = batch.strokes?.[j] ?? batch.stroke;
-  return stroke === null || stroke === undefined ? themeInk : resolve(stroke);
-}
-
-/**
- * Draw segments with Θ(runs) stroke() calls: mono batches (no per-segment
- * `strokes`) are one path; per-segment colors collapse contiguous same-color
- * runs. Optional `includes` skips primitives for focus-mask subset passes.
- */
-function drawSegments(
-  ctx: CanvasRenderingContext2D,
-  batch: SegmentsBatch,
-  theme: ThemeTokens,
-  resolve: ColorResolver,
-  includes?: (index: number) => boolean,
-): void {
-  const themeInk = resolve(themeVar("ink", theme));
-  ctx.lineWidth = batch.linewidth;
-  const n = batch.segments.length / 4;
-  if (n === 0) return;
-
-  if (batch.strokes === undefined) {
-    ctx.strokeStyle = segmentStrokeAt(batch, 0, themeInk, resolve);
-    ctx.beginPath();
-    let traced = false;
-    for (let j = 0; j < n; j++) {
-      if (includes !== undefined && !includes(j)) continue;
-      traceSegment(ctx, batch, j);
-      traced = true;
-    }
-    if (traced) ctx.stroke();
-    return;
-  }
-
-  let runStart = 0;
-  while (runStart < n) {
-    if (includes !== undefined && !includes(runStart)) {
-      runStart++;
-      continue;
-    }
-    const color = segmentStrokeAt(batch, runStart, themeInk, resolve);
-    let runEnd = runStart + 1;
-    while (runEnd < n && segmentStrokeAt(batch, runEnd, themeInk, resolve) === color) runEnd++;
-    ctx.strokeStyle = color;
-    ctx.beginPath();
-    let traced = false;
-    for (let j = runStart; j < runEnd; j++) {
-      if (includes !== undefined && !includes(j)) continue;
-      traceSegment(ctx, batch, j);
-      traced = true;
-    }
-    if (traced) ctx.stroke();
-    runStart = runEnd;
-  }
-}
-
-function drawBatchInner(
-  ctx: CanvasRenderingContext2D,
-  batch: GeometryBatch,
-  theme: ThemeTokens,
-  resolve: ColorResolver,
-): void {
-  switch (batch.kind) {
-    case "points":
-      drawPoints(ctx, batch, theme, resolve);
-      break;
-    case "paths":
-      drawPaths(ctx, batch, theme, resolve);
-      break;
-    case "rects": {
-      const themeFill = resolve(themeVar(batch.fillRole ?? "accent", theme));
-      const n = batch.rects.length / 4;
-      for (let j = 0; j < n; j++) {
-        const fill = batch.fills?.[j] ?? batch.fill;
-        ctx.fillStyle = fill === null || fill === undefined ? themeFill : resolve(fill);
-        ctx.fillRect(
-          batch.rects[j * 4]!,
-          batch.rects[j * 4 + 1]!,
-          batch.rects[j * 4 + 2]!,
-          batch.rects[j * 4 + 3]!,
-        );
-        if (batch.stroke !== undefined) {
-          ctx.strokeStyle = resolve(batch.stroke ?? themeVar("ink", theme));
-          ctx.lineWidth = batch.strokeWidth ?? 1;
-          ctx.strokeRect(
-            batch.rects[j * 4]!,
-            batch.rects[j * 4 + 1]!,
-            batch.rects[j * 4 + 2]!,
-            batch.rects[j * 4 + 3]!,
-          );
-        }
-      }
-      break;
-    }
-    case "segments":
-      drawSegments(ctx, batch, theme, resolve);
-      break;
-    case "glyphs":
-      // Text always renders as SVG (module docs); nothing to draw.
-      break;
-  }
-}
-
-function drawBatchSubsetInner(
-  ctx: CanvasRenderingContext2D,
-  batch: GeometryBatch,
-  theme: ThemeTokens,
-  resolve: ColorResolver,
-  mask: PrimitiveFocusMask,
-  focused: boolean,
-): void {
-  const includes = (index: number) => maskIncludes(mask, index) === focused;
-  switch (batch.kind) {
-    case "points":
-      drawPointsSubset(ctx, batch, theme, resolve, mask, focused);
-      break;
-    case "paths":
-      drawPathsSubset(ctx, batch, theme, resolve, mask, focused);
-      break;
-    case "rects": {
-      const themeFill = resolve(themeVar(batch.fillRole ?? "accent", theme));
-      const n = batch.rects.length / 4;
-      for (let j = 0; j < n; j++) {
-        if (!includes(j)) continue;
-        const fill = batch.fills?.[j] ?? batch.fill;
-        ctx.fillStyle = fill === null || fill === undefined ? themeFill : resolve(fill);
-        ctx.fillRect(
-          batch.rects[j * 4]!,
-          batch.rects[j * 4 + 1]!,
-          batch.rects[j * 4 + 2]!,
-          batch.rects[j * 4 + 3]!,
-        );
-        if (batch.stroke !== undefined) {
-          ctx.strokeStyle = resolve(batch.stroke ?? themeVar("ink", theme));
-          ctx.lineWidth = batch.strokeWidth ?? 1;
-          ctx.strokeRect(
-            batch.rects[j * 4]!,
-            batch.rects[j * 4 + 1]!,
-            batch.rects[j * 4 + 2]!,
-            batch.rects[j * 4 + 3]!,
-          );
-        }
-      }
-      break;
-    }
-    case "segments":
-      drawSegments(ctx, batch, theme, resolve, includes);
-      break;
-    case "glyphs":
-      // Text always renders as SVG; the mask still addresses glyph primitives
-      // so SVG and canvas callers can share one renderer-neutral mask shape.
-      break;
-  }
-}
-
-function batchPrimitiveCount(batch: GeometryBatch): number {
-  switch (batch.kind) {
-    case "points":
-    case "rects":
-    case "segments":
-    case "glyphs":
-      return batch.rowIndex.length;
-    case "paths":
-      return Math.max(0, batch.pathOffsets.length - 1);
-    default:
-      return 0;
-  }
-}
-
-function maskIsAllFocused(mask: PrimitiveFocusMask, count: number): boolean {
-  if (count === 0) return true;
-  if ("isFocused" in mask && mask.primitiveCount === count)
-    return mask.focusedCount === mask.primitiveCount;
-  for (let index = 0; index < count; index++) {
-    if (!maskIncludes(mask, index)) return false;
-  }
-  return true;
-}
-
-function drawBatchSubset(
-  ctx: CanvasRenderingContext2D,
-  batch: GeometryBatch,
-  theme: ThemeTokens,
-  resolve: ColorResolver,
-  mask: PrimitiveFocusMask,
-  focused: boolean,
-  alphaMultiplier: number,
-): void {
-  const alpha = ctx.globalAlpha;
-  ctx.globalAlpha = alpha * batch.alpha * alphaMultiplier;
-  drawBatchSubsetInner(ctx, batch, theme, resolve, mask, focused);
-  ctx.globalAlpha = alpha;
-}
-
-/** Draw one batch in panel-local coordinates (alpha applied per batch). */
-export function drawBatch(
-  ctx: CanvasRenderingContext2D,
-  batch: GeometryBatch,
-  theme: ThemeTokens,
-  resolve: ColorResolver,
-): void {
-  const alpha = ctx.globalAlpha;
-  ctx.globalAlpha = alpha * batch.alpha;
-  drawBatchInner(ctx, batch, theme, resolve);
-  ctx.globalAlpha = alpha;
-}
+export type { ColorResolver } from "./canvas-dom.js";
+export { cssColorResolver, drawClippedToPanel, sizeCanvasForDpr } from "./canvas-dom.js";
+export type { CanvasFocusPresentation, PrimitiveFocusMask } from "./canvas-marks.js";
+export { drawBatch } from "./canvas-marks.js";
 
 /**
  * Group geometry batches by panel once (issue #185): O(B) instead of
